Lower accessibility standards :(
This seems to be a bug with Lighthouse.

It states that a <div> can't live inside a <dl>, however the HTML5 spec
does allow this.

For now, as the result is exactly 0.9, it's ok to lower the test to
check for >=. Ideally the bug will get fixed in Lighthouse soon, and we
can raise to >0.9 again